update help command & ___ manage -> configure ___ `actions manage` is now `configure actions`. `reminders manage` is now `configure reminders`
